Better random article selection 4 manual analysis
Moved manual analysis code into the analysis folder. I'm not sure if it will stay here or go back to the tidying section. It's not "cleaning" any of the articles themselves, but it is moving them around, which feels like it could be "tidying" to me.

New strategy only grabs a few new articles at a time. The reason is I'm still getting more articles as time goes by, so I want to continue to give new articles a chance to get randomly selected. Previously, the articles I were analyzing were biased towards OA articles because that's all I had at the time. This is NOT an ideal randomization procedure, but given that I need to wait around to get articles from publishers with more stringent requirements, I'm taking that hit because it allows me to work on other aspects of the project.