Manual screening by journal name
I used my existing list of eligible articles to find which journals had a high number of INeligible articles. Then I searched unpaywall_info.csv for the corresponding journals. In short, I was able to find a good deal more articles that needed to be classified as non-human, and a few as not op-rr.