Create clean_txt_files.py
For some reason I'm not converting several thousand txt files...I may not keep this script.

The reason for creating the script was to reduce the time it took to read in txt files. By creating the "tidy" versions of the txt files I hoped to save on computation time when doing any later analysis.