Add Author Offprint copy to gitignore
Per IJSM email: "Attached please find an electronic reprint (PDF file) for your records. Please note that this PDF was created especially for you: it is intended for your personal use as well as for the professional exchange with colleagues. Please feel free to upload the file on your personal homepage."